requirements.txt
Flask-JWT-Extended
Flask-Restful
Flask-SQLAlchemy